requirements.txt
fdfgen
sh
pathlib
mysql-connector-python
PyMySQL
flask